requirements.txt
streamlit
streamlit-webrtc
Pillow
pillow-heif
opencv-python-headless
pyzbar
streamlit-image-coordinates
uuid
pathlib
sqlalchemy
requests
python-dotenv
zipfile38
numpy
